Export access to the TextBoxController
I've been trying to work on some slightly more complicated efforts in my code editor and realized that I really, really want access to the *TextBoxController field in the mixins.TextBox.  While I wait to hear back about whether or not I'm using the right approach, I'm going to use a fork of gxui that includes a getter for the controller field.