Fix Forward Delets in TextBoxController
A previous change made the delete key practically useless, as the cursor would move with each delete.
This was due to the 'At' field of TextBoxEdits referring to the starting location for positive edits,
but the ending location for negative edits.  Now, updateSelectionsForEdits pays attention to the
starting position of the edit, rather than the 'At' value.