Use list-like architecture for driver events
A deep enough tree architecture was causing a nasty bug to crop up where the number of pending calls
in the driver's pendingApp event channel was the size of the channel's buffer, and further calls to
driver.Call() within a function that was running on the UI goroutine would cause the UI to hang.

This updates the data structure to be a queue with a channel that only holds the on-deck call.  All
calls after the one on-deck are kept in a linked list style data structure.